+++ examples/examples/stream.rs
@@ -34,5 +34,5 @@ async fn main() {
         AnyConfig::default(),
     ));
 
-    elfo::start(topology).await;
+    elfo::init::start(topology).await;
 }